+++ modules/adlooxRtdProvider.js
@@ -25,7 +25,6 @@ import {
   deepAccess,
   deepClone,
   deepSetValue,
-  getGptSlotInfoForAdUnitCode,
   isArray,
   isBoolean,
   isInteger,
@@ -37,6 +36,7 @@ import {
   parseUrl,
   safeJSONParse
 } from '../src/utils.js';
+import {getGptSlotInfoForAdUnitCode} from '../libraries/gptUtils/gptUtils.js';
 
 const MODULE_NAME = 'adloox';
 const MODULE = `${MODULE_NAME}RtdProvider`;
