+++ test/spec/modules/sonobiBidAdapter_spec.js
@@ -1,9 +1,10 @@
-import { expect } from 'chai'
-import { spec, _getPlatform } from 'modules/sonobiBidAdapter.js'
-import { newBidder } from 'src/adapters/bidderFactory.js'
-import { userSync } from '../../../src/userSync.js';
-import { config } from 'src/config.js';
+import {expect} from 'chai';
+import {_getPlatform, spec} from 'modules/sonobiBidAdapter.js';
+import {newBidder} from 'src/adapters/bidderFactory.js';
+import {userSync} from '../../../src/userSync.js';
+import {config} from 'src/config.js';
 import * as utils from '../../../src/utils.js';
+import {getGptSlotInfoForAdUnitCode} from '../../../libraries/gptUtils/gptUtils.js';
 
 describe('SonobiBidAdapter', function () {
   const adapter = newBidder(spec)
@@ -254,7 +255,7 @@ describe('SonobiBidAdapter', function () {
     });
     afterEach(function () {
       userSync.canBidderRegisterSync.restore();
-      utils.getGptSlotInfoForAdUnitCode.restore();
+      getGptSlotInfoForAdUnitCode.restore();
       sandbox.restore();
     });
     let bidRequest = [{
